paper-cite-agent/requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
typer>=0.12.0
pyyaml>=6.0
requests>=2.31.0
lxml>=5.0.0
# 可选：更精准的语义相似度（安装后自动启用）
# sentence-transformers>=3.0.0
